chore(capacitor): remove init generator
