--- apps/docs/docs/capacitor/generators/init.md
@@ -1,22 +1,0 @@
----
-id: init
-title: Init
-sidebar_label: Init
----
-
-Initializes the `@nxtend/capacitor` plugin.
-
-This generator adds the following Capacitor dependencies:
-
-```
-@capacitor/core
-@capacitor/cli
-@capacitor/android
-@capacitor/ios
-```
-
-## Usage
-
-```
-nx generate @nxtend/capacitor:init
-```
