+++ packages/capacitor/generators.json
@@ -3,13 +3,6 @@
   "name": "NxtendCapacitor",
   "version": "0.0.1",
   "generators": {
-    "init": {
-      "factory": "./src/generators/init/generator",
-      "schema": "./src/generators/init/schema.json",
-      "description": "Initialize the @nxtend/capacitor plugin",
-      "aliases": ["ng-add"],
-      "hidden": true
-    },
     "capacitor-project": {
       "factory": "./src/generators/capacitor-project/generator",
       "schema": "./src/generators/capacitor-project/schema.json",
