+++ packages/capacitor/src/generators/capacitor-project/generator.ts
@@ -1,7 +1,6 @@
-import { GeneratorCallback, Tree } from '@nrwl/devkit';
-import { runTasksInSerial } from '@nrwl/workspace/src/utilities/run-tasks-in-serial';
-import init from '../init/generator';
+import { Tree } from '@nrwl/devkit';
 import { addCapacitorConfig } from './lib/add-capacitor-config';
+import { addDependencies } from './lib/add-dependencies';
 import { addProject } from './lib/add-project';
 import { normalizeOptions } from './lib/normalize-options';
 import { updateProjectGitignore } from './lib/update-project-gitignore';
@@ -12,15 +11,13 @@ export async function capacitorProjectGenerator(
   host: Tree,
   options: CapacitorGeneratorSchema
 ) {
-  const tasks: GeneratorCallback[] = [];
   const normalizedOptions = normalizeOptions(host, options);
-  const initTask = await init(host);
-  tasks.push(initTask);
+  const installTask = addDependencies(host);
   addCapacitorConfig(host, normalizedOptions);
   updateProjectPackageJson(host, normalizedOptions);
   updateProjectGitignore(host, normalizedOptions);
   addProject(host, normalizedOptions);
-  return runTasksInSerial(...tasks);
+  return installTask;
 }
 
 export default capacitorProjectGenerator;
