+++ packages/capacitor/src/generators/capacitor-project/lib/add-dependencies.ts
