--- packages/capacitor/src/generators/init/schema.json
@@ -1,9 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema",
-  "cli": "nx",
-  "id": "Init",
-  "title": "Add Nxtend Capacitor Schematics",
-  "type": "object",
-  "properties": {},
-  "required": []
-}
